+++ models/contacts.json
@@ -55,8 +55,8 @@
   },
   {
     "id": "rsKkOQUi80UsgVPCcLZZW",
-    "name": "qwdqwdqwd",
-    "email": "nulla.ante@vestibul.co.uk",
-    "phone": "(992) 914-3eeee33"
+    "name": "Alec Howard",
+    "email": "Donec.elementum@scelil",
+    "phone": "(992) 914-3723"
   }
 ]
