+++ models/copy/get.routers.js
@@ -1,0 +1,36 @@
+// /* eslint-disable spaced-comment */
+// const express = require('express')
+
+// const fs = require("fs");
+// const path = require("path");
+// const router = express.Router()
+
+
+// // const contacts1 = require("../models/contacts.json")
+
+// const contactsPath = path.resolve("./models/contacts.json");
+// const data = fs.readFileSync(contactsPath);
+// const contacts = JSON.parse(data);
+
+
+// function listContacts() {
+//   try {
+//     return contacts;
+// } catch (err) {
+//   console.log(err.message);
+//   }
+// }
+
+// const getById = (contactId) => {
+//   return contacts.find((contact) => contact.id === contactId);
+// }
+
+// /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+
+
+
+
+
+// //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+// module.exports = router
