+++ package.json
@@ -13,7 +13,7 @@
     "cross-env": "7.0.3",
     "express": "4.17.1",
     "morgan": "1.10.0",
-    "nanoid": "^5.0.4"
+    "nanoid": "^3.0.0"
   },
   "devDependencies": {
     "eslint": "7.19.0",
