--- routes/patch.routers.js
@@ -1,8 +1,0 @@
-const express = require('express')
-
-const router = express.Router()
-
-router.patch('/', async (req, res, next) => {
-  res.json({ message: 'PATCH' })
-})
-module.exports = router
